cleaned up script
